recover some deprecated
