+++ utils/lr_scheduler.py
@@ -1,5 +1,3 @@
-
-
 def adjust_learning_rate(optimizer, epoch, cfgs):
     """Sets the learning rate to the initial LR decayed by 10 every 30 epochs"""
     lr = cfgs['lr'] * (0.1 ** (epoch // 30))
