+++ utils/miscellaneous.py
@@ -1,0 +1,30 @@
+import errno
+import os
+import torch
+import shutil
+import PIL
+from torch.utils.collect_env import get_pretty_env_info
+
+
+def mkdir(path):
+    try:
+        os.makedirs(path)
+    except OSError as e:
+        if e.errno != errno.EEXIST:
+            raise
+
+
+def get_pil_version():
+    return "\n        Pillow ({})".format(PIL.__version__)
+
+
+def collect_env_info():
+    env_str = get_pretty_env_info()
+    env_str += get_pil_version()
+    return env_str
+
+
+def save_checkpoint(state, is_best, filename='checkpoint.pth.tar'):
+    torch.save(state, filename + '_latest.pth.tar')
+    if is_best:
+        shutil.copyfile(filename + '_latest.pth.tar', filename + '_best.pth.tar')
